Update jquery.kinetic.js
This added scroller delegation with the x and y setting. You can choose any string selector as text when initializing kinetic under the x or y option instead of just true and false. _getScroller now accepts an axis variable to determine which selector and setting we are choosing.

window, body and html automatically choose $(window). Anything else will choose $(this.settings.y) or $(this.settings.x) respectively.

This is necessary so that boxes that go across screen on mobile can still move the page up and down as you can target y with window, html or body and it will delegate the movement to the window itself and not get stuck in the box.